+++ africanus/averaging/time_and_channel_mapping.py
@@ -2,54 +2,241 @@
 
 
 from collections import namedtuple
+from numbers import Number
 
 import numpy as np
 import numba
+from numba.experimental import jitclass
 
 from africanus.averaging.support import unique_time, unique_baselines
-from africanus.util.numba import is_numba_type_none, generated_jit, njit, jit
+from africanus.util.numba import is_numba_type_none, generated_jit, njit
 
 
 class RowMapperError(Exception):
     pass
 
 
-def is_flagged_factory(have_flag_row):
-    if have_flag_row:
-        def impl(flag_row, r):
-            return flag_row[r] != 0
+def _numba_type(obj):
+    if isinstance(obj, np.ndarray):
+        return numba.typeof(obj.dtype).dtype
+    elif isinstance(obj, numba.types.npytypes.Array):
+        return obj.dtype
+    elif isinstance(obj, (np.dtype, numba.types.Type)):
+        return numba.typeof(obj).dtype
+    elif isinstance(obj, Number):
+        return numba.typeof(obj)
     else:
-        def impl(flag_row, r):
-            return False
-
-    return njit(nogil=True, cache=True)(impl)
+        raise TypeError(f"Unhandled type {type(obj)}")
 
 
-def output_factory(have_flag_row):
-    if have_flag_row:
-        def impl(rows, flag_row):
-            return np.zeros(rows, dtype=flag_row.dtype)
+def binner_factory(time, interval, antenna1, antenna2,
+                   flag_row, time_bin_secs):
+    if flag_row is None:
+        have_flag_row = False
     else:
-        def impl(rows, flag_row):
-            return None
+        have_flag_row = not is_numba_type_none(flag_row)
 
-    return njit(nogil=True, cache=True)(impl)
+    class Binner:
+        def __init__(self, time, interval, antenna1, antenna2,
+                     flag_row, time_bin_secs):
+            ubl, _, bl_inv, _ = unique_baselines(antenna1, antenna2)
+            utime, _, time_inv, _ = unique_time(time)
 
+            ntime = utime.shape[0]
+            nbl = ubl.shape[0]
+            self.bl_inv = bl_inv
+            self.time_inv = time_inv
+            self.out_rows = 0
+            row_lookup = np.full((nbl, ntime), -1, dtype=np.intp)
 
-def set_flag_row_factory(have_flag_row):
+            # Create a mapping from the full bl x time resolution back
+            # to the original input rows
+            for r, (t, bl) in enumerate(zip(time_inv, bl_inv)):
+                if row_lookup[bl, t] == -1:
+                    row_lookup[bl, t] = r
+                else:
+                    raise ValueError("Duplicate (TIME, ANTENNA1, ANTENNA2) "
+                                     "combinations were discovered in the "
+                                     "input data. This is usually caused by "
+                                     "not partitioning your data sufficiently "
+                                     "by indexing columns, DATA_DESC_ID "
+                                     "and SCAN_NUMBER in particular.")
+
+            sentinel = np.finfo(time.dtype).max
+
+            self.row_lookup = row_lookup
+            self.time_bin_secs = time_bin_secs
+            self.time_lookup = np.full((nbl, ntime), sentinel, time.dtype)
+            self.interval_lookup = np.zeros((nbl, ntime), interval.dtype)
+            self.bin_flagged = np.full((nbl, ntime), False)
+            self.bin_lookup = np.full((nbl, ntime), -1)
+
+            self.time = time
+            self.interval = interval
+
+            if have_flag_row:
+                self.flag_row = flag_row
+
+        def start_baseline(self):
+            self.tbin = 0
+            self.rs = 0
+            self.re = 0
+            self.bin_count = 0
+            self.bin_flag_count = 0
+
+        def finalise_baseline(self):
+            self.out_rows += self.tbin
+
+        @property
+        def bin_empty(self):
+            return self.bin_count == 0
+
+        def start_bin(self, row, flag_row):
+            self.rs = row
+            self.re = row
+            self.bin_count = 1
+            self.bin_flag_count = int(have_flag_row and flag_row[row] != 0)
+
+        def add_row(self, row, time, interval, flag_row):
+            rs = self.rs
+
+            if self.re == row:
+                raise ValueError("start_bin should be called "
+                                 "to start a bin before add_row "
+                                 "is called.")
+
+            dt = ((time[row] + 0.5*interval[row]) -
+                  (time[rs] - 0.5*interval[rs]))
+
+            if dt > self.time_bin_secs:
+                return False
+            self.re = row
+            self.bin_count += 1
+            flagged = have_flag_row and flag_row[row] != 0
+            self.bin_flag_count += int(flagged)
+
+            return True
+
+        def finalise_bin(self, bl, next_row, time, interval):
+            rs = self.rs
+            re = self.re
+            tbin = self.tbin
+
+            # No interpolation required
+            if rs == re:
+                bin_time = time[rs]
+                bin_interval = interval[rs]
+            else:
+                # Interpolate between bin start and end times.
+
+                # 1. We use the previous bin and the first row of
+                #    the next bin to establish this where possible
+                #    as these points determine the full bin extent.
+                # 2. Otherwise we must use the first and last row
+                #    in the bin for the first and last bin
+                #    respectively. This is not as accurate as (1),
+                #    but is best effort in the case of missing edge data
+                if tbin > 0:
+                    # Use the bin prior to this one to establish
+                    # the start time of the bin
+                    time_start = (self.time_lookup[bl, tbin - 1] +
+                                  0.5*self.interval_lookup[bl, tbin - 1])
+                else:
+                    # Use the time and interval of the starting row to
+                    # establish the start time of the bin
+                    time_start = time[rs] - 0.5*interval[rs]
+
+                # Find bin ending time
+                if next_row != re:
+                    # Use the time and interval of the next row outside
+                    # the bin to establish the end time of the bin
+                    time_end = time[next_row] - 0.5*interval[next_row]
+
+                    if time_end - time_start > self.time_bin_secs:
+                        time_end = time_start + self.time_bin_secs
+                else:
+                    # Use the time and interval of the ending row
+                    # to establish the end time of the bin
+                    time_end = time[re] + 0.5*interval[re]
+
+                # Establish the midpoint
+                bin_time = 0.5*(time_start + time_end)
+                bin_interval = time_end - time_start
+
+            self.time_lookup[bl, tbin] = bin_time
+            self.interval_lookup[bl, tbin] = bin_interval
+            self.bin_flagged[bl, tbin] = self.bin_count == self.bin_flag_count
+
+            self.tbin += 1
+
+        def execute(self):
+            row_lookup = self.row_lookup
+            time = self.time
+            interval = self.interval
+            flag_row = self.flag_row if have_flag_row else None
+            bin_lookup = self.bin_lookup
+
+            # Average times over each baseline and construct the
+            # bin_lookup and time_lookup arrays
+            for bl in range(row_lookup.shape[0]):
+                self.start_baseline()
+
+                for t in range(row_lookup.shape[1]):
+                    r = row_lookup[bl, t]
+
+                    if r == -1:
+                        continue
+
+                    if self.bin_empty:
+                        self.start_bin(r, flag_row)
+                    elif not self.add_row(r, time, interval, flag_row):
+                        # Can't add a new row to this bin, close it
+                        # and start a new one
+                        self.finalise_bin(bl, r, time, interval)
+                        self.start_bin(r, flag_row)
+
+                    # Register the output time bin for this row
+                    bin_lookup[bl, t] = self.tbin
+
+                # Close any open bins
+                if not self.bin_empty:
+                    self.finalise_bin(bl, r, time, interval)
+
+                self.finalise_baseline()
+
+    time = _numba_type(time)
+    interval = _numba_type(interval)
+    antenna1 = _numba_type(antenna1)
+    antenna2 = _numba_type(antenna2)
+    time_bin_secs = _numba_type(time_bin_secs)
+
+    spec = [
+        ('out_rows', numba.uintp),
+        ('rs', numba.intp),
+        ('re', numba.intp),
+        ('tbin', numba.intp),
+        ('bin_count', numba.uintp),
+        ('bin_flag_count', numba.uintp),
+        ('bl_inv', numba.uintp[:]),
+        ('time_inv', numba.uintp[:])]
+
+    spec.extend([
+        ('time_lookup', time[:, :]),
+        ('interval_lookup', interval[:, :]),
+        ('row_lookup', numba.intp[:, :]),
+        ('bin_lookup', numba.intp[:, :]),
+        ('bin_flagged', numba.bool_[:, :]),
+        ('time_bin_secs', time_bin_secs)])
+
+    spec.extend([
+        ('time', time[:]),
+        ('interval', interval[:])])
+
     if have_flag_row:
-        def impl(flag_row, in_row, out_flag_row, out_row, flagged):
-            if flag_row[in_row] == 0 and flagged:
-                raise RowMapperError("Unflagged input row contributing "
-                                     "to flagged output row. "
-                                     "This should never happen!")
+        flag_row = _numba_type(flag_row)
+        spec.append(('flag_row', flag_row[:]))
 
-            out_flag_row[out_row] = (1 if flagged else 0)
-    else:
-        def impl(flag_row, in_row, out_flag_row, out_row, flagged):
-            pass
-
-    return njit(nogil=True, cache=True)(impl)
+    return jitclass(spec)(Binner)
 
 
 RowMapOutput = namedtuple("RowMapOutput",
@@ -179,133 +366,47 @@ def row_mapper(time, interval, antenna1, antenna2,
 
     """
     have_flag_row = not is_numba_type_none(flag_row)
-    is_flagged_fn = is_flagged_factory(have_flag_row)
+    have_time_bin_secs = not is_numba_type_none(time_bin_secs)
+    time_bin_secs_type = time_bin_secs if have_time_bin_secs else time.dtype
 
-    output_flag_row = output_factory(have_flag_row)
-    set_flag_row = set_flag_row_factory(have_flag_row)
+    JitBinner = binner_factory(time, interval, antenna1, antenna2,
+                               flag_row, time_bin_secs_type)
 
     def impl(time, interval, antenna1, antenna2,
              flag_row=None, time_bin_secs=1):
-        ubl, _, bl_inv, _ = unique_baselines(antenna1, antenna2)
-        utime, _, time_inv, _ = unique_time(time)
+        # If we don't have time_bin_secs
+        # set it to the maximum floating point value,
+        # effectively ignoring this limit
+        if not have_time_bin_secs:
+            time_bin_secs = np.finfo(time.dtype).max
 
-        nbl = ubl.shape[0]
-        ntime = utime.shape[0]
-
-        sentinel = np.finfo(time.dtype).max
-        out_rows = numba.uint32(0)
-
-        scratch = np.full(3*nbl*ntime, -1, dtype=np.int32)
-        row_lookup = scratch[:nbl*ntime].reshape(nbl, ntime)
-        bin_lookup = scratch[nbl*ntime:2*nbl*ntime].reshape(nbl, ntime)
-        inv_argsort = scratch[2*nbl*ntime:]
-        time_lookup = np.zeros((nbl, ntime), dtype=time.dtype)
-        interval_lookup = np.zeros((nbl, ntime), dtype=interval.dtype)
-
-        # Is the entire bin flagged?
-        bin_flagged = np.zeros((nbl, ntime), dtype=np.bool_)
-
-        # Create a mapping from the full bl x time resolution back
-        # to the original input rows
-        for r in range(time.shape[0]):
-            bl = bl_inv[r]
-            t = time_inv[r]
-
-            if row_lookup[bl, t] == -1:
-                row_lookup[bl, t] = r
-            else:
-                raise ValueError("Duplicate (TIME, ANTENNA1, ANTENNA2) "
-                                 "combinations were discovered in the input "
-                                 "data. This is usually caused by not "
-                                 "partitioning your data sufficiently "
-                                 "by indexing columns, DATA_DESC_ID "
-                                 "and SCAN_NUMBER in particular.")
-
-        # Average times over each baseline and construct the
-        # bin_lookup and time_lookup arrays
-        for bl in range(ubl.shape[0]):
-            tbin = numba.int32(0)
-            bin_count = numba.int32(0)
-            bin_flag_count = numba.int32(0)
-            bin_low = time.dtype.type(0)
-
-            for t in range(utime.shape[0]):
-                # Lookup input row
-                r = row_lookup[bl, t]
-
-                # Ignore if not present
-                if r == -1:
-                    continue
-
-                # At this point, we decide whether to contribute to
-                # the current bin, or create a new one. We don't add
-                # the current sample to the current bin if
-                # high - low >= time_bin_secs
-                half_int = interval[r] * 0.5
-
-                # We're starting a new bin anyway,
-                # just set the lower bin value
-                if bin_count == 0:
-                    bin_low = time[r] - half_int
-                # If we exceed the seconds in the bin,
-                # normalise the time and start a new bin
-                elif time[r] + half_int - bin_low > time_bin_secs:
-                    # Normalise and flag the bin
-                    # if total counts match flagged counts
-                    if bin_count > 0:
-                        time_lookup[bl, tbin] /= bin_count
-                        bin_flagged[bl, tbin] = bin_count == bin_flag_count
-                    # There was nothing in the bin
-                    else:
-                        time_lookup[bl, tbin] = sentinel
-                        bin_flagged[bl, tbin] = False
-
-                    tbin += 1
-                    bin_count = 0
-                    bin_low = time[r] - half_int
-                    bin_flag_count = 0
-
-                # Record the output bin associated with the row
-                bin_lookup[bl, t] = tbin
-
-                # Time + Interval take unflagged + unflagged
-                # samples into account (nominal value)
-                time_lookup[bl, tbin] += time[r]
-                interval_lookup[bl, tbin] += interval[r]
-                bin_count += 1
-
-                # Record flags
-                if is_flagged_fn(flag_row, r):
-                    bin_flag_count += 1
-
-            # Normalise the last bin if it has entries in it
-            if bin_count > 0:
-                time_lookup[bl, tbin] /= bin_count
-                bin_flagged[bl, tbin] = bin_count == bin_flag_count
-                tbin += 1
-
-            # Add this baseline's number of bins to the output rows
-            out_rows += tbin
-
-            # Set any remaining bins to sentinel value and unflagged
-            for b in range(tbin, ntime):
-                time_lookup[bl, b] = sentinel
-                bin_flagged[bl, b] = False
+        binner = JitBinner(time, interval, antenna1, antenna2,
+                           flag_row, time_bin_secs)
+        binner.execute()
 
         # Flatten the time lookup and argsort it
-        flat_time = time_lookup.ravel()
-        flat_int = interval_lookup.ravel()
+        flat_time = binner.time_lookup.ravel()
+        flat_int = binner.interval_lookup.ravel()
         argsort = np.argsort(flat_time, kind='mergesort')
+        inv_argsort = np.empty_like(argsort)
 
         # Generate lookup from flattened (bl, time) to output row
         for i, a in enumerate(argsort):
             inv_argsort[a] = i
 
         # Construct the final row map
-        row_map = np.empty((time.shape[0]), dtype=np.uint32)
+        row_map = np.empty(time.shape[0], dtype=np.uint32)
+
+        nbl, ntime = binner.row_lookup.shape
+        out_rows = binner.out_rows
+        bin_lookup = binner.bin_lookup
+        bin_flagged = binner.bin_flagged
+        bl_inv = binner.bl_inv
+        time_inv = binner.time_inv
 
         # Construct output flag row, if necessary
-        out_flag_row = output_flag_row(out_rows, flag_row)
+        out_flag_row = (np.zeros(out_rows, dtype=flag_row.dtype)
+                        if have_flag_row else None)
 
         # foreach input row
         for in_row in range(time.shape[0]):
@@ -321,10 +422,14 @@ def row_mapper(time, interval, antenna1, antenna2,
             if out_row >= out_rows:
                 raise RowMapperError("out_row >= out_rows")
 
-            # Handle output row flagging
-            set_flag_row(flag_row, in_row,
-                         out_flag_row, out_row,
-                         bin_flagged[bl, tbin])
+            if have_flag_row:
+                flagged = bin_flagged[bl, tbin]
+                if flag_row[in_row] == 0 and flagged:
+                    raise RowMapperError("Unflagged input row contributing "
+                                         "to flagged output row. "
+                                         "This should never happen!")
+
+                out_flag_row[out_row] = 1 if flagged else 0
 
             row_map[in_row] = out_row
 
@@ -336,7 +441,7 @@ def row_mapper(time, interval, antenna1, antenna2,
     return impl
 
 
-@jit(nopython=True, nogil=True, cache=True)
+@njit(nogil=True, cache=True)
 def channel_mapper(nchan, chan_bin_size=1):
     chan_map = np.empty(nchan, dtype=np.uint32)